scripts/yokai-print-server/requirements.txt
supabase>=2.3.0
python-escpos>=3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
